+++ packages/core/src/bases/chain-wallet.ts
@@ -1,22 +1,14 @@
 /* eslint-disable no-unused-expressions */
-import {
-  CosmWasmClient,
-  SigningCosmWasmClient,
-  SigningCosmWasmClientOptions,
-} from '@cosmjs/cosmwasm-stargate';
+import { SigningCosmWasmClientOptions } from '@cosmjs/cosmwasm-stargate';
 import { EncodeObject, OfflineSigner } from '@cosmjs/proto-signing';
 import {
-  calculateFee,
   GasPrice,
-  SigningStargateClient,
   SigningStargateClientOptions,
-  StargateClient,
   StargateClientOptions,
   StdFee,
 } from '@cosmjs/stargate';
-import { TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
+import type { TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
 
-import { NameService } from '../name-service';
 import {
   ChainRecord,
   CosmosClientType,
@@ -28,13 +20,7 @@ import {
   State,
   Wallet,
 } from '../types';
-import {
-  getFastestEndpoint,
-  getIsLazy,
-  getNameServiceRegistryFromChainName,
-  isValidEndpoint,
-} from '../utils';
-import { MainWalletBase } from './main-wallet';
+import type { MainWalletBase } from './main-wallet';
 import { WalletBase } from './wallet';
 
 export class ChainWalletBase extends WalletBase {
@@ -234,6 +220,7 @@ export class ChainWalletBase extends WalletBase {
   getRpcEndpoint = async (
     isLazy?: boolean
   ): Promise<string | ExtendedHttpEndpoint> => {
+    const { getIsLazy } = await import('../utils');
     const lazy = getIsLazy(
       void 0,
       this.isLazy,
@@ -252,6 +239,7 @@ export class ChainWalletBase extends WalletBase {
 
     const nodeType = 'rpc';
 
+    const { isValidEndpoint } = await import('../utils');
     if (
       this._rpcEndpoint &&
       (await isValidEndpoint(this._rpcEndpoint, nodeType, lazy, this.logger))
@@ -259,6 +247,7 @@ export class ChainWalletBase extends WalletBase {
       return this._rpcEndpoint;
     }
 
+    const { getFastestEndpoint } = await import('../utils');
     this._rpcEndpoint = await getFastestEndpoint(
       this.rpcEndpoints,
       nodeType,
@@ -270,6 +259,7 @@ export class ChainWalletBase extends WalletBase {
   getRestEndpoint = async (
     isLazy?: boolean
   ): Promise<string | ExtendedHttpEndpoint> => {
+    const { getIsLazy } = await import('../utils');
     const lazy = getIsLazy(
       void 0,
       this.isLazy,
@@ -288,6 +278,7 @@ export class ChainWalletBase extends WalletBase {
 
     const nodeType = 'rest';
 
+    const { isValidEndpoint } = await import('../utils');
     if (
       this._restEndpoint &&
       (await isValidEndpoint(this._restEndpoint, nodeType, lazy, this.logger))
@@ -295,6 +286,7 @@ export class ChainWalletBase extends WalletBase {
       return this._restEndpoint;
     }
 
+    const { getFastestEndpoint } = await import('../utils');
     this._restEndpoint = await getFastestEndpoint(
       this.restEndpoints,
       nodeType,
@@ -303,19 +295,23 @@ export class ChainWalletBase extends WalletBase {
     return this._restEndpoint;
   };
 
-  getStargateClient = async (): Promise<StargateClient> => {
+  getStargateClient = async () => {
     const rpcEndpoint = await this.getRpcEndpoint();
+    const { StargateClient } = await import('@cosmjs/stargate');
     return StargateClient.connect(rpcEndpoint, this.stargateOptions);
   };
 
-  getCosmWasmClient = async (): Promise<CosmWasmClient> => {
+  getCosmWasmClient = async () => {
     const rpcEndpoint = await this.getRpcEndpoint();
+    const { CosmWasmClient } = await import('@cosmjs/cosmwasm-stargate');
     return CosmWasmClient.connect(rpcEndpoint);
   };
 
-  getNameService = async (): Promise<NameService> => {
+  getNameService = async () => {
     const client = await this.getCosmWasmClient();
+    const { getNameServiceRegistryFromChainName } = await import('../utils');
     const registry = getNameServiceRegistryFromChainName(this.chainName);
+    const { NameService } = await import('../name-service');
     return new NameService(client, registry);
   };
 
@@ -330,12 +326,14 @@ export class ChainWalletBase extends WalletBase {
     );
   }
 
-  getSigningStargateClient = async (): Promise<SigningStargateClient> => {
+  getSigningStargateClient = async () => {
     const rpcEndpoint = await this.getRpcEndpoint();
 
     if (!this.offlineSigner) {
       await this.initOfflineSigner();
     }
+
+    const { SigningStargateClient } = await import('@cosmjs/stargate');
 
     return SigningStargateClient.connectWithSigner(
       rpcEndpoint,
@@ -344,13 +342,14 @@ export class ChainWalletBase extends WalletBase {
     );
   };
 
-  getSigningCosmWasmClient = async (): Promise<SigningCosmWasmClient> => {
+  getSigningCosmWasmClient = async () => {
     const rpcEndpoint = await this.getRpcEndpoint();
 
     if (!this.offlineSigner) {
       await this.initOfflineSigner();
     }
 
+    const { SigningCosmWasmClient } = await import('@cosmjs/cosmwasm-stargate');
     return SigningCosmWasmClient.connectWithSigner(
       rpcEndpoint,
       this.offlineSigner,
@@ -401,8 +400,9 @@ export class ChainWalletBase extends WalletBase {
     }
     const client = await this.getSigningClient(type);
     const gasEstimation = await client.simulate(this.address, messages, memo);
+    const { calculateFee } = await import('@cosmjs/stargate');
     return calculateFee(
-      Math.round(gasEstimation * (multiplier || 1.3)),
+      Math.round(gasEstimation * (multiplier || 1.4)),
       gasPrice
     );
   };
@@ -431,6 +431,7 @@ export class ChainWalletBase extends WalletBase {
 
   broadcast = async (signedMessages: TxRaw, type?: CosmosClientType) => {
     const client = await this.getSigningClient(type);
+    const { TxRaw } = await import('cosmjs-types/cosmos/tx/v1beta1/tx');
     const txBytes = TxRaw.encode(signedMessages).finish();
 
     let timeoutMs: number | undefined, pollIntervalMs: number | undefined;