+++ packages/core/src/types/view.ts
@@ -1,6 +1,7 @@
 import { ChainWalletBase } from '../bases';
 import { WalletRepo } from '../repository';
 import { Dispatch } from './common';
+import { ModalOptions } from './wallet';
 
 export enum ModalView {
   WalletList = 'WalletList',
@@ -22,6 +23,7 @@ export interface WalletViewProps {
   onClose: () => void;
   onReturn: () => void;
   wallet: ChainWalletBase;
+  options?: ModalOptions;
 }
 
 interface RefObject<T> {