+++ packages/core/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.5](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.8.4...@cosmos-kit/core@2.8.5) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/core
+
+
+
+
+
+## [2.8.4](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.8.3...@cosmos-kit/core@2.8.4) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/core
+
+
+
+
+
+## [2.8.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.8.2...@cosmos-kit/core@2.8.3) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/core
+
+
+
+
+
+## [2.8.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.8.1...@cosmos-kit/core@2.8.2) (2024-01-15)
+
+### Bug Fixes
+
+- wrong error throw ([9adffac](https://github.com/cosmology-tech/cosmos-kit/commit/9adffac4c86b9169e0664f880dd916f1107d934b))
+
+## [2.8.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.8.0...@cosmos-kit/core@2.8.1) (2024-01-08)
+
+**Note:** Version bump only for package @cosmos-kit/core
+
+# [2.8.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.7.14...@cosmos-kit/core@2.8.0) (2024-01-07)
+
+**Note:** Version bump only for package @cosmos-kit/core
+
 ## [2.7.14](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.7.13...@cosmos-kit/core@2.7.14) (2023-12-26)
 
 **Note:** Version bump only for package @cosmos-kit/core