+++ packages/example/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.5](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.11.4...example@2.11.5) (2024-01-19)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [2.11.4](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.11.3...example@2.11.4) (2024-01-19)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [2.11.3](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.11.2...example@2.11.3) (2024-01-19)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [2.11.2](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.11.1...example@2.11.2) (2024-01-15)
+
+**Note:** Version bump only for package example
+
+## [2.11.1](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.11.0...example@2.11.1) (2024-01-08)
+
+**Note:** Version bump only for package example
+
+# [2.11.0](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.10.6...example@2.11.0) (2024-01-07)
+
+**Note:** Version bump only for package example
+
 ## [2.10.6](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.10.5...example@2.10.6) (2023-12-28)
 
 **Note:** Version bump only for package example