+++ packages/react-lite/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.5](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.6.4...@cosmos-kit/react-lite@2.6.5) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/react-lite
+
+
+
+
+
+## [2.6.4](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.6.3...@cosmos-kit/react-lite@2.6.4) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/react-lite
+
+
+
+
+
+## [2.6.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.6.2...@cosmos-kit/react-lite@2.6.3) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/react-lite
+
+
+
+
+
+## [2.6.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.6.1...@cosmos-kit/react-lite@2.6.2) (2024-01-15)
+
+**Note:** Version bump only for package @cosmos-kit/react-lite
+
+## [2.6.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.6.0...@cosmos-kit/react-lite@2.6.1) (2024-01-08)
+
+**Note:** Version bump only for package @cosmos-kit/react-lite
+
+# [2.6.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.5.17...@cosmos-kit/react-lite@2.6.0) (2024-01-07)
+
+**Note:** Version bump only for package @cosmos-kit/react-lite
+
 ## [2.5.17](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.5.16...@cosmos-kit/react-lite@2.5.17) (2023-12-28)
 
 **Note:** Version bump only for package @cosmos-kit/react-lite