+++ wallets/cosmos-extension-metamask/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.5](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.4.4...@cosmos-kit/cosmos-extension-metamask@0.4.5) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+
+
+
+
+## [0.4.4](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.4.3...@cosmos-kit/cosmos-extension-metamask@0.4.4) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+
+
+
+
+## [0.4.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.4.2...@cosmos-kit/cosmos-extension-metamask@0.4.3) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+
+
+
+
+## [0.4.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.4.1...@cosmos-kit/cosmos-extension-metamask@0.4.2) (2024-01-15)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+## [0.4.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.4.0...@cosmos-kit/cosmos-extension-metamask@0.4.1) (2024-01-08)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+# [0.4.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.3.4...@cosmos-kit/cosmos-extension-metamask@0.4.0) (2024-01-07)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
 ## [0.3.4](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.3.3...@cosmos-kit/cosmos-extension-metamask@0.3.4) (2023-12-26)
 
 **Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask