+++ wallets/initia/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.5](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/initia@0.2.4...@cosmos-kit/initia@0.2.5) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/initia
+
+
+
+
+
+## [0.2.4](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/initia@0.2.3...@cosmos-kit/initia@0.2.4) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/initia
+
+
+
+
+
+## [0.2.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/initia@0.2.2...@cosmos-kit/initia@0.2.3) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/initia
+
+
+
+
+
+## [0.2.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/initia@0.2.1...@cosmos-kit/initia@0.2.2) (2024-01-15)
+
+**Note:** Version bump only for package @cosmos-kit/initia
+
+## [0.2.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/initia@0.2.0...@cosmos-kit/initia@0.2.1) (2024-01-08)
+
+**Note:** Version bump only for package @cosmos-kit/initia
+
+# [0.2.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/initia@0.1.0...@cosmos-kit/initia@0.2.0) (2024-01-07)
+
+**Note:** Version bump only for package @cosmos-kit/initia
+
 # 0.1.0 (2023-12-26)
 
 ### Bug Fixes