+++ wallets/keplr-extension/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.5](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.7.4...@cosmos-kit/keplr-extension@2.7.5) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-extension
+
+
+
+
+
+## [2.7.4](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.7.3...@cosmos-kit/keplr-extension@2.7.4) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-extension
+
+
+
+
+
+## [2.7.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.7.2...@cosmos-kit/keplr-extension@2.7.3) (2024-01-19)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-extension
+
+
+
+
+
+## [2.7.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.7.1...@cosmos-kit/keplr-extension@2.7.2) (2024-01-15)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-extension
+
+## [2.7.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.7.0...@cosmos-kit/keplr-extension@2.7.1) (2024-01-08)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-extension
+
+# [2.7.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.6.4...@cosmos-kit/keplr-extension@2.7.0) (2024-01-07)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-extension
+
 ## [2.6.4](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.6.3...@cosmos-kit/keplr-extension@2.6.4) (2023-12-26)
 
 **Note:** Version bump only for package @cosmos-kit/keplr-extension